cleanup and smaller fixes for the player
