+++ pages/api/calculateColor.ts
