--- pages/api/rebuilt/presenceCheck.ts
@@ -1,18 +1,0 @@
-/* query the database to check if the url is already present */
-import { NextApiRequest, NextApiResponse } from 'next'
-import { prisma } from './prismaClientProvider';
-
-export default async function handle(req: NextApiRequest, res: NextApiResponse) {
-
-    const track = await prisma.track.findFirst({
-        where: {
-            url: req.body.url,
-        },
-    });
-
-    if (track) {
-        res.json({ status: true});
-    } else {
-        res.json({ status: false });
-    }
-}
