--- pages/api/tagging.ts
@@ -1,27 +1,0 @@
-import { Configuration, OpenAIApi } from "openai";
-import { NextApiRequest, NextApiResponse } from "next";
-
-const configuration = new Configuration({
-    apiKey: process.env.OPENAI_API_KEY,
-});
-
-const openai = new OpenAIApi(configuration);
-
-const tagging = async (req: NextApiRequest, res: NextApiResponse) => {
-
-    const { title, artist } = req.body;
-
-    const gptResponse = await
-                openai.createChatCompletion({
-                    model: "gpt-4",
-                    messages: [
-                        { "role": "system", "content": "You are a helpful assistant that likes to tag youtube video titles. Example: 'rock, pop, alternative, indie, alternative rock'. Refrain from using more than one word per tag and seperate the tags by ','. Do not add special characters to tags like # or @. Translate tags to english." },
-                        { "role": "user", "content": "Please create a raw string without formatting of 5 tags for the given track title: '" + title + "' by '" + artist + "'. Try to categorize the tracks into exclusively one of the following: 'ambience', 'music' or 'standalone'(music & ambience). Try to use tags that describe the scene which the track portrays" },
-                    ]
-                })
-            const response = gptResponse.data.choices[0].message?.content;
-
-    res.status(200).json(response);
-}
-
-export default tagging;
